+++ express/blocks/pricing/pricing.js
@@ -15,10 +15,10 @@ import {
   addPublishDependencies,
   createTag,
   getHelixEnv,
-  getOffer,
   getIcon,
 // eslint-disable-next-line import/no-unresolved
 } from '../../scripts/scripts.js';
+import { getOffer } from '../../scripts/utils/pricing.js';
 
 function replaceUrlParam(url, paramName, paramValue) {
   const params = url.searchParams;
